Feat/service test

+++ src/main/java/com/fundingproject/funding/service/FundingOpportunityService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com.fundingproject.funding.entity.FundingOpportunity;
 import com.fundingproject.funding.enums.FundingStatus;
+import com.fundingproject.funding.exception.FundingOpportunityServiceException;
 import com.fundingproject.funding.repository.FundingOpportunityRepository;
 
 @Service
@@ -20,9 +21,16 @@ public class FundingOpportunityService {
     this.fundingOpportunityRepository = fundingOpportunityRepository;
   }
 
+
   public List<FundingOpportunity> getAllFundingOpportunities() {
-    return fundingOpportunityRepository.findAll();
-  }
+        try {
+            return fundingOpportunityRepository.findAll();
+        } catch (Exception e) {
+            throw new FundingOpportunityServiceException(
+                "Error fetching funding opportunities from database", e
+            );
+        }
+    }
 
   public FundingOpportunity getFundingOpportunityById(UUID id) throws NoSuchElementException {
     return fundingOpportunityRepository.findById(id)
@@ -58,8 +66,11 @@ public class FundingOpportunityService {
   }
 
   public void deleteFundingOpportunity(UUID id) {
+    if (!fundingOpportunityRepository.existsById(id)) {
+        throw new NoSuchElementException("Funding opportunity with id " + id + " not found");
+    }
     fundingOpportunityRepository.deleteById(id);
-  }
+}
 
   public List<FundingOpportunity> getByFundingStatus(FundingStatus fundingStatus) {
     return fundingOpportunityRepository.findByFundingStatus(fundingStatus);
